test(ui): split the Details and port tests into their own file
Details rendering is width-sensitive and its port fields carry the ownership
rules, so both are checked together.

Tests move unchanged.
